+++ app/ninja/repositories/PaymentRepository.php
@@ -18,7 +18,7 @@ class PaymentRepository
                     ->where('payments.account_id', '=', \Auth::user()->account_id)
                     ->where('clients.deleted_at', '=', null)
                     ->where('contacts.is_primary', '=', true)   
-                    ->select('payments.public_id', 'payments.transaction_reference', 'clients.name as client_name', 'clients.public_id as client_public_id', 'payments.amount', 'payments.payment_date', 'invoices.public_id as invoice_public_id', 'invoices.invoice_number', 'clients.currency_id', 'contacts.first_name', 'contacts.last_name', 'contacts.email', 'payment_types.name as payment_type', 'payments.account_gateway_id');        
+                    ->select('payments.public_id', 'payments.transaction_reference', 'clients.name as client_name', 'clients.public_id as client_public_id', 'payments.amount', 'payments.payment_date', 'invoices.public_id as invoice_public_id', 'invoices.invoice_number', 'clients.currency_id', 'contacts.first_name', 'contacts.last_name', 'contacts.email', 'payment_types.name as payment_type', 'payments.account_gateway_id', 'payments.deleted_at', 'payments.is_deleted');
 
         if (!\Session::get('show_trash:payment'))
         {
@@ -148,13 +148,20 @@ class PaymentRepository
 
         foreach ($payments as $payment) 
         {            
-            if ($action == 'delete') 
+            if ($action == 'restore')
             {
-                $payment->is_deleted = true;
-                $payment->save();
-            } 
+                $payment->restore();
+            }
+            else
+            {            
+                if ($action == 'delete') 
+                {
+                    $payment->is_deleted = true;
+                    $payment->save();
+                } 
 
-            $payment->delete();
+                $payment->delete();
+            }
         }
 	
 		return count($payments);
